Adding a full table fetch.
In some circumstances where bertconf still is the mechanism
used but the config data needs further treatment, it should be possible
to fetch entries without knowing the key.

This patch adds a new function to fetch the entire content of a
table, batched by chunks of 500 to avoid locking up a scheduler
for too long on copying data. A single list with all entries is
returned to the user.

The value is not configurable for now -- if there is a need for it,
the code can be extended to allow it.