Adding versioning functions
This patch adds two versionning functions, described in the
README.md file. They allow to kind of 'poll' for new versions
of configuration on a given table, without needing to read
the entire table to know if there is new material.

This is especially useful when using the all/1 function to do
further processing somewhere else based on some configuration
data.